requirements.txt
scipy==1.10.1
numpy==1.24.3
openpyxl
excel2img
PyYAML